moved GA version of modem.Data to legacy
