+++ examples/create_modem_input.py
@@ -28,10 +28,10 @@ import sys
 import matplotlib.pyplot as plt
 import numpy as np
 
+from legacy.modeling.modem_data import Data
 from mtpy.core.edi_collection import EdiCollection
 from mtpy.gui.SmartMT.utils.matplotlib_utils import gen_hist_bins
 from mtpy.modeling.modem_covariance import Covariance
-from mtpy.modeling.modem_data import Data
 from mtpy.modeling.modem_model import Model
