+++ legacy/modeling/modem_data.py
@@ -195,7 +195,7 @@ class Data(object):
     :Example 1 --> create inversion period list: ::
 
         >>> import os
-        >>> from mtpy.modeling.modem_data import Data
+        >>> from legacy.modeling.modem_data import Data
         >>> edi_path = r"/home/mt/edi_files"
         >>> edi_list = [os.path.join(edi_path, edi) \
                         for edi in os.listdir(edi_path)\
@@ -207,7 +207,7 @@ class Data(object):
     :Example 2 --> set inverions period list from data: ::
 
         >>> import os
-        >>> from mtpy.modeling.modem_data import Data
+        >>> from legacy.modeling.modem_data import Data
         >>> edi_path = r"/home/mt/edi_files"
         >>> edi_list = [os.path.join(edi_path, edi) \
                         for edi in os.listdir(edi_path)\
@@ -223,7 +223,7 @@ class Data(object):
 
     :Example 3 --> change error values: ::
 
-        >>> from mtpy.modeling.modem_data import Data
+        >>> from legacy.modeling.modem_data import Data
         >>> mdr = Data()
         >>> mdr.read_data_file(r"/home/modem/inv1/ModEM_Data.dat")
         >>> mdr.error_type = 'floor'
@@ -233,7 +233,7 @@ class Data(object):
 
     :Example 4 --> change inversion type: ::
 
-        >>> from mtpy.modeling.modem_data import Data
+        >>> from legacy.modeling.modem_data import Data
         >>> mdr = Data()
         >>> mdr.read_data_file(r"/home/modem/inv1/ModEM_Data.dat")
         >>> mdr.inv_mode = '3'
@@ -1247,7 +1247,7 @@ class Data(object):
 
         :Example: ::
 
-            >>> from mtpy.modeling.modem_data import Data as modem
+            >>> from legacy.modeling.modem_data import Data as modem
             >>> mdr = modem.Data()
             >>> mdr.convert_ws3dinv_data_file(r"/home/ws3dinv/inv1/WSData.dat",
                     station_fn=r"/home/ws3dinv/inv1/WS_Station_Locations.txt")
